% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_latest_vic_gtfs.R
\name{download_latest_vic_gtfs}
\alias{download_latest_vic_gtfs}
\title{Download the latest Victorian GTFS schedule}
\usage{
download_latest_vic_gtfs(to = "gtfs")
}
\arguments{
\item{to}{Directory where the GTFS files should be saved.}
}
\value{
The path to the folder where the GTFS files were extracted.
}
\description{
Downloads the most recent GTFS schedule from the Victorian Government
website and saves it to a folder for further extraction.
}
